Move low-level http functions to a raw module
Move the low-level http functions that work with serde maps directly into an
`http::raw` module, and re-export them from the `http` module for
backwards compatibility purposes.
